requirements.txt
absl-py @ file:///C:/ci/absl-py_1603893176799/work
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828549974/work
astor==0.8.1
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1605538148575/work
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gast @ file:///tmp/build/80754af9/gast_1597433534803/work
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1597406403308/work
h5py==2.10.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython==5.8.0
ipython-genutils==0.2.0
Jinja2==2.11.2
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras==2.2.4
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
Markdown @ file:///C:/ci/markdown_1605111187600/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1605115881283/work
notebook @ file:///C:/ci/notebook_1601494955742/work
numpy==1.16.4
olefile==0.46
packaging==20.4
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow @ file:///C:/ci/pillow_1603821929285/work
prometheus-client==0.8.0
prompt-toolkit==1.0.15
protobuf==3.13.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pygame==2.0.1
Pygments @ file:///tmp/build/80754af9/pygments_1604103097372/work
pyparsing==2.4.7
pyreadline==2.1
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
pyserial @ file:///tmp/build/80754af9/pyserial_1594419199743/work
python-dateutil==2.8.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
scipy @ file:///C:/ci/scipy_1597686737426/work
Send2Trash==1.5.0
simplegeneric==0.8.1
six @ file:///C:/ci/six_1605205426665/work
tensorboard==1.14.0
tensorflow @ file:///D:/bld/tensorflow_1594833538462/work/tensorflow-1.14.0-cp37-cp37m-win_amd64.whl
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
